feat(daffio): update marketing pages
