+++ apps/daffio/src/app/content/home/components/home-callout-commerce/home-callout-commerce.component.spec.ts
@@ -8,30 +8,24 @@ import {
   TestBed,
 } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { provideRouter } from '@angular/router';
 
-import { DaffioHomeCalloutPwaComponent } from './home-callout-pwa.component';
+import { DaffioHomeCalloutCommerceComponent } from './home-callout-commerce.component';
 
 @Component({
-  template: `<daffio-home-callout-pwa></daffio-home-callout-pwa>`,
-  imports: [DaffioHomeCalloutPwaComponent],
+  template: `<daffio-home-callout-commerce></daffio-home-callout-commerce>`,
+  imports: [DaffioHomeCalloutCommerceComponent],
 })
 class WrapperComponent {}
 
-describe('DaffioHomeCalloutPwaComponent', () => {
+describe('DaffioHomeCalloutCommerceComponent', () => {
   let wrapper: WrapperComponent;
-  let component: DaffioHomeCalloutPwaComponent;
+  let component: DaffioHomeCalloutCommerceComponent;
   let de: DebugElement;
   let fixture: ComponentFixture<WrapperComponent>;
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      imports: [
-        WrapperComponent,
-      ],
-      providers: [
-        provideRouter([]),
-      ],
+      imports: [WrapperComponent],
     })
       .compileComponents();
   }));
@@ -39,7 +33,7 @@ describe('DaffioHomeCalloutPwaComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(WrapperComponent);
     wrapper = fixture.debugElement.componentInstance;
-    de = fixture.debugElement.query(By.css('daffio-home-callout-pwa'));
+    de = fixture.debugElement.query(By.css('daffio-home-callout-commerce'));
     component = de.componentInstance;
     fixture.detectChanges();
   });
@@ -48,9 +42,9 @@ describe('DaffioHomeCalloutPwaComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should add a class of "daffio-home-callout-pwa" to the host element', () => {
+  it('should add a class of "daffio-home-callout-commerce" to the host element', () => {
     expect(de.classes).toEqual(jasmine.objectContaining({
-      'daffio-home-callout-pwa': true,
+      'daffio-home-callout-commerce': true,
     }));
   });
 });
