+++ apps/daffio/src/app/content/home/components/home-callout-platforms/home-callout-platforms.component.html
@@ -1,7 +1,7 @@
 <daff-callout textAlignment="center">
   <daff-container size="md">
     <div daffCalloutTitle>
-      Seamless multi-platform ecommerce integration
+      Seamless multi-platform ecommerce integration.
     </div>
     <p daffCalloutSubtitle>
       Transition between Magento, Shopify or any other platforms without the hassle of reinventing your front-end.
