--- apps/daffio/src/app/content/home/components/home-callout-pwa/home-callout-pwa.component.html
@@ -1,41 +1,0 @@
-<daff-callout textAlignment="center" color="theme">
-  <daff-container size="md">
-    <h3 daffCalloutTitle>
-      Build the next generation of ecommerce stores
-    </h3>
-    <p daffCalloutSubtitle>
-      Deliver high quality user experiences by leveraging the benefits of Progressive Web Apps.
-    </p>
-    <div daffCalloutBody class="daffio-home-callout-pwa__body">
-      <div class="daffio-home-callout-pwa__card-grid">
-        <daff-card>
-          <div daffCardTitle>Offline Support</div>
-          <div daffCardContent>
-            <p>Take advantage of service workers to cache key assets and data on user devices. Empower your store and allow users to shop independent of network quality.</p>
-          </div>
-        </daff-card>
-        <daff-card>
-          <div daffCardTitle>Native Functionalities</div>
-          <div daffCardContent>
-            <p>Utilize native device features like home screen access, push notifications, and immersive full screen experiences without the need of an app store.</p>
-          </div>
-        </daff-card>
-        <daff-card>
-          <div daffCardTitle>Discoverable</div>
-          <div daffCardContent>
-            <p>Get all the benefits of a seamless mobile experience without sacrificing SEO. Just like a typical website, all content is indexable and searchable.</p>
-          </div>
-        </daff-card>
-        <daff-card>
-          <div daffCardTitle>Fast</div>
-          <div daffCardContent>
-            <p>PWAs are light. They exponentially improve loading time, creating a smooth native performance for users.</p>
-          </div>
-        </daff-card>
-      </div>
-      <a routerLink="/why-pwa" daff-button color="primary" class="daffio-home-callout-pwa__button">
-        Learn more about PWAs
-      </a>
-    </div>
-  </daff-container>
-</daff-callout>
