--- apps/daffio/src/app/content/home/components/home-callout-pwa/home-callout-pwa.component.scss
@@ -1,55 +1,0 @@
-@use 'utilities' as daff;
-
-.daffio-home-callout-pwa {
-	&__body {
-		display: grid;
-		grid-template-columns: 1fr;
-		grid-template-areas:
-			'card-grid'
-			'button';
-		grid-gap: 48px;
-	}
-
-	&__card-grid {
-		grid-area: card-grid;
-		display: grid;
-		grid-template-columns: 1fr;
-		grid-template-areas:
-			'a'
-			'b'
-			'c'
-			'd';
-		grid-gap: 16px;
-
-		@include daff.breakpoint(tablet) {
-			grid-template-columns: 1fr 1fr;
-			grid-template-areas:
-				'a b'
-				'c d';
-			grid-gap: 16px;
-		}
-
-		> * {
-			&:first-child {
-				grid-area: a;
-			}
-
-			&:nth-child(2) {
-				grid-area: b;
-			}
-
-			&:nth-child(3) {
-				grid-area: c;
-			}
-
-			&:last-child {
-				grid-area: d;
-			}
-		}
-	}
-
-	&__button {
-		grid-area: button;
-		justify-self: center;
-	}
-}
