+++ apps/daffio/src/app/content/home/components/home-hero/home-hero.component.scss
@@ -12,7 +12,7 @@
 	background-size: cover;
 
 	@include daff.breakpoint(big-tablet) {
-		height: 80vh;
+		height: 70vh;
 	}
 
 	&__actions {
@@ -24,23 +24,5 @@
 
 	&__subtitle-focus {
 		font-weight: 600;
-	}
-
-	&__code-copy {
-		@include daff.clickable();
-		appearance: none;
-		backdrop-filter: blur(0.5rem);
-		display: flex;
-		align-items: center;
-		justify-content: center;
-		gap: 0.75rem;
-		font-size: daff.$font-size-sm;
-		border-radius: 2rem;
-		margin: 3rem auto 0;
-		padding: 0.5rem 1rem;
-
-		@include daff.breakpoint(tablet) {
-			padding: 1rem 1.5rem;
-		}
 	}
 }
