+++ apps/daffio/src/app/content/home/view/home-view.component.html
@@ -1,4 +1,4 @@
 <daffio-home-hero></daffio-home-hero>
-<daffio-home-callout-pwa></daffio-home-callout-pwa>
+<daffio-home-callout-commerce></daffio-home-callout-commerce>
 <daffio-home-callout-platforms></daffio-home-callout-platforms>
 <daffio-home-callout-sponsors></daffio-home-callout-sponsors>
