+++ apps/daffio/src/app/core/footer/footer/footer.component.html
@@ -5,6 +5,7 @@
         <a routerLink="{{ link.path }}" class="daffio-footer__link">{{ link.title }}</a>
       }
       <a class="daffio-footer__link" href="https://github.com/sponsors/graycoreio" target="_blank" rel="noopener noreferrer">Sponsor</a>
+      <a class="daffio-footer__link" href="https://github.com/graycoreio/daffodil/blob/develop/CONTRIBUTING.md" target="_blank" rel="noopener noreferrer">Contribute</a>
     </div>
     <a class="daffio-footer__logo" routerLink="/">
       <daff-branding-logo type="icon"></daff-branding-logo>
